Prepare wox.py to accept multiple inputs over stdin
Streaming over standard input enables a plugin to process multiple JsonRPC
messages without starting up a new instance of python each time which is
slow. Improves python plugin responsiveness especially with plugins that
have a lot of setup as this can be retained between calls.

All message now return at least an empty line to ensure wox doesn't deadlock.

As it still may be useful for debugging the old way of sending messages
as an argument is retained.