optimize/program/query_match: Optimize matching and cancellation in Program plugin
It's easy to cause high CPU usage in "Program" plugin, ususally when
matching a long query string (e.g. "aaaaaaaaaaaaaaaaaaaaaaaaaaa")
against a very long program description string, because of the nature
of matcher function's recursion and the usage of parallelism. Meanwhile,
when input speed is really fast, the lack of timely cancellation of this
procedure brings terrible nearly 100% CPU usage, usually making Wox
stuck.

This commit introduces cancallation in query process in Program
plugin as much as possible. In addition and especially, the introduction
of cancellation token in `StringMatcher` object.