optimize/wox/result_selected: Keep selected index to 0 in result listbox, when new results arrive
When Wox updates the results frequently (likely because different
plugins have different response time), the currently selected result
index (`SelectedIndex`) in result listbox may change, like from 0
to 2, to keep the same item be selected in a new result list
with different item order. This is not we want, because when
we type something and hit Enter, we expect it's always the
first item (aka index 0) that will be opened.

How to keep selected index to 0 in result querybox seems to be a tough
problem. Most of the time you set the newResults, update the bound
collection, notify the change, and it will take several milliseconds for
the framework to respond and set `SelectedIndex` to a different value
than 0. This fix uses a `CollectionChangedPrioritized` event and some
delayed tasks to set the index back to 0, hoping to mitigate this issue.